update maskImg & maskBackground

See merge request s06-webmobile1-sub2/S06P12E202!178

+++ back-end/src/main/java/com/ssafy/api/request/ProfileModifyMaskPostReq.java
@@ -13,5 +13,5 @@ import lombok.Setter;
 @ApiModel("ProfileModifyMaskPostRequest")
 public class ProfileModifyMaskPostReq {
 	@ApiModelProperty(name="값", example="mask_id")
-	Long mask_id;
+	int mask_id;
 }
